Upstream version 2.3.0.1

# gpg: Signature made Παρ 23 Μάρ 2018 11:22:47 μμ EET
# gpg:                using RSA key 3E02FD6656295952110BAB99F51B18C720248224
# gpg:                issuer "apoikos@debian.org"
# gpg: Good signature from "Apollon Oikonomopoulos <apoikos@dmesg.gr>" [ultimate]
# gpg:                 aka "Apollon Oikonomopoulos <apoikos@gmail.com>" [ultimate]
# gpg:                 aka "Apollon Oikonomopoulos <apoikos@debian.org>" [ultimate]

+++ src/pop3-login/pop3-proxy.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2004-2018 Dovecot authors, see the included COPYING file */
+/* Copyright (c) 2004-2017 Dovecot authors, see the included COPYING file */
 
 #include "login-common.h"
 #include "ioloop.h"
@@ -30,7 +30,7 @@ static int proxy_send_login(struct pop3_client *client, struct ostream *output)
 {
 	struct dsasl_client_settings sasl_set;
 	const unsigned char *sasl_output;
-	unsigned int len;
+	size_t len;
 	const char *mech_name, *error;
 	string_t *str = t_str_new(128);
 
@@ -111,7 +111,7 @@ pop3_proxy_continue_sasl_auth(struct client *client, struct ostream *output,
 {
 	string_t *str;
 	const unsigned char *data;
-	unsigned int data_len;
+	size_t data_len;
 	const char *error;
 	int ret;
